dev-requirements.txt
flake8    # for the travis build
black
pytest
pytest-ckan